move fluctuations in core
