+++ athena/client/fetch.py
@@ -6,7 +6,7 @@ from tqdm import tqdm
 
 from athena.client.binance import BinanceClient
 from athena.core.dataset_layout import DatasetLayout
-from athena.core.interfaces.fluctuations import Fluctuations, load_candles_from_file
+from athena.core.fluctuations import Fluctuations, load_candles_from_file
 from athena.core.market_entities import Candle
 from athena.core.types import Coin, Period
