+++ athena/entrypoints/backtest.py
@@ -2,9 +2,10 @@ from pathlib import Path
 
 import click
 
-from athena.core.config import DataConfig, StrategyConfig
+from athena.configs import DataConfig, StrategyConfig
 from athena.settings import Settings
-from athena.core.interfaces import DatasetLayout, Fluctuations
+from athena.core.fluctuations import Fluctuations
+from athena.core.dataset_layout import DatasetLayout
 from athena.entrypoints.utils import load_config
 from athena.performance.report import build_and_save_trading_report
 from athena.performance.trading_session import TradingSession
