+++ athena/performance/report.py
@@ -6,7 +6,7 @@ import plotly.graph_objects as go
 from jinja2 import Environment, PackageLoader, select_autoescape
 from plotly.subplots import make_subplots
 
-from athena.core.interfaces import Fluctuations
+from athena.core.fluctuations import Fluctuations
 from athena.core.market_entities import Portfolio, Position
 from athena.core.types import Coin
 from athena.performance.models import (
