+++ athena/performance/trading_session.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, ConfigDict
 
-from athena.core.config import DataConfig
-from athena.core.interfaces.fluctuations import Fluctuations
+from athena.configs import DataConfig
+from athena.core.fluctuations import Fluctuations
 from athena.core.market_entities import Portfolio, Position, Trade, Candle
 from athena.core.types import Signal
 from athena.tradingtools.strategies.strategy import Strategy
