+++ athena/tradingtools/indicators/momentum/rsi.py
@@ -1,6 +1,6 @@
 from ta.momentum import RSIIndicator, StochRSIIndicator
 
-from athena.core.interfaces import Fluctuations
+from athena.core.fluctuations import Fluctuations
 from athena.tradingtools.indicators.common import IndicatorLine
