+++ athena/tradingtools/indicators/trend/ichimoku.py
@@ -1,7 +1,7 @@
 from pydantic import Field
 from ta.trend import IchimokuIndicator
 
-from athena.core.interfaces import Fluctuations
+from athena.core.fluctuations import Fluctuations
 from athena.tradingtools.indicators.common import IndicatorLine
