+++ athena/tradingtools/indicators/trend/moving_average.py
@@ -1,7 +1,7 @@
 from pydantic import Field
 from ta.trend import EMAIndicator, SMAIndicator
 
-from athena.core.interfaces import Fluctuations
+from athena.core.fluctuations import Fluctuations
 from athena.tradingtools.indicators.common import IndicatorLine
