+++ athena/tradingtools/strategies/strategy.py
@@ -1,7 +1,7 @@
 import re
 from typing import Iterable
 
-from athena.core.interfaces.fluctuations import Fluctuations
+from athena.core.fluctuations import Fluctuations
 from athena.core.market_entities import Candle
 from athena.core.types import Signal
