+++ tests/core/interfaces/test_fluctuations_functions.py
@@ -4,7 +4,7 @@ import logging
 import pytest
 
 from athena.core.market_entities import Candle
-from athena.core.interfaces.fluctuations import (
+from athena.core.fluctuations import (
     _convert_candles_to_period,
     _merge_candles,
     _sanitize_candles,
