+++ tests/entrypoints/test_backtest.py
@@ -5,7 +5,8 @@ import pytest
 from click.testing import CliRunner
 
 from athena.cli import app
-from athena.core.interfaces import DatasetLayout, Fluctuations
+from athena.core.fluctuations import Fluctuations
+from athena.core.dataset_layout import DatasetLayout
 from athena.core.types import Coin, Period
