+++ tests/entrypoints/test_download.py
@@ -3,7 +3,8 @@ import datetime
 from click.testing import CliRunner
 
 from athena.cli import app
-from athena.core.interfaces import DatasetLayout, Fluctuations
+from athena.core.fluctuations import Fluctuations
+from athena.core.dataset_layout import DatasetLayout
 from athena.core.types import Coin, Period
